move atomics test and wait_async_mock.js to js-sys; empty futures test file
